+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FEATURES:
 
+* **New Data Source:** `proxmox_virtual_environment_datastores`
 * **New Data Source:** `proxmox_virtual_environment_group`
 * **New Data Source:** `proxmox_virtual_environment_groups`
 * **New Data Source:** `proxmox_virtual_environment_nodes`
@@ -12,7 +13,6 @@ FEATURES:
 * **New Data Source:** `proxmox_virtual_environment_user`
 * **New Data Source:** `proxmox_virtual_environment_users`
 * **New Data Source:** `proxmox_virtual_environment_version`
-
 * **New Resource:** `proxmox_virtual_environment_group`
 * **New Resource:** `proxmox_virtual_environment_pool`
 * **New Resource:** `proxmox_virtual_environment_role`